вынести радар в отдельный класс
